+++ extensions/data-plane/data-plane-http-oauth2-core/src/main/java/org/eclipse/edc/connector/dataplane/http/oauth2/Oauth2HttpRequestParamsDecorator.java
@@ -50,5 +50,4 @@ public class Oauth2HttpRequestParamsDecorator implements HttpParamsDecorator {
             return params;
         }
     }
-
 }