+++ CHANGELOG.md
@@ -6,11 +6,28 @@
 
 #### Changes
 
+#### Details
+
+#### Compatibility
+
+#### Resolution plan
+
+---
+
+## [0.11.1.2] - UNRELEASED
+
+### Overview
+
+Add OAuth tokens caching.
+
+#### Changes
+
 - Add caching for data plane OAuth requests
 
 #### Details
 
 - Add `TokenCache`ing in `Oauth2HttpRequestParamsDecorator`.
+  - The minimum time to live for the token (how long before it expires it should be renewed) can be configured with `sovity.edc.dataplane.oauth2.cache.minimum.time.to.live`.
 
 #### Compatibility
 
@@ -18,8 +35,9 @@ No breaking expected.
 
 #### Resolution plan
 
----
+None.
 
+---
 
 ## [0.11.1.1] - 2025-04-25